Category delete only if not in use

+++ app/Http/Controllers/V1/Category/DeleteController.php
@@ -4,13 +4,19 @@ namespace App\Http\Controllers\V1\Category;
 
 use App\Http\Controllers\Controller;
 use App\Models\V1\Category;
+use App\Models\V1\Item;
 use Illuminate\Http\Request;
 
 class DeleteController extends Controller
 {
     public function __invoke(Category $category)
     {
-        $category->delete();
-        return response()->json(['message' => 'Category deleted'], 200);
+        if ($category->items()->exists()) {
+            $message = 'Can not delete. Category is in use.';
+        } else {
+            $category->delete();
+            $message = 'Category deleted successfully.';
+        }
+        return response()->json(['message' => $message], 200);
     }
 }
